CataBlog Release 1.6
* New: WYSIWYG MCE Editor for the catalog item description.
* New: Edit category name and slug.
* New: Added French localization. (Thanks Nicolas)
* Fix: Javascript syntax error fixed in Remove All CataBlog Data script.
* Fix: Forwards browser to the CataBlog interface when a library item is loaded into the standard WP Post editor.
* Fix: Edit library item panels do not overlap on small screens.
* Fix: Category list line up correctly in IE now.
* Fix: Text input boxes shouldn't cut off dangling letters in IE.
* Fix: Modal window's position is now fixed, meaning no scrolling back to the top of the page to use it.



git-svn-id: https://plugins.svn.wordpress.org/catablog/trunk@524478 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ templates/admin-remove.php
@@ -24,23 +24,23 @@
 		
 		var progress_bar = $('#catablog-progress div.catablog-progress-bar');
 		
-
+		if (!confirm("Are you sure you want to remove all CataBlog data from your WordPress site? You cannot undo this.")) {
+			return false;
+		}
+		
 		
 		delete_library_part();
-		
-		
-		
 		
 		function delete_library_part() {
 			var params  = {
 				'action':   'catablog_delete_library',
-				'security': '<?php echo wp_create_nonce("catablog-delete-library") ?>',
+				'security': '<?php echo wp_create_nonce("catablog-delete-library") ?>'
 			}
 
 
 			$.post(ajaxurl, params, function(data) {
 				try {
-					var json = eval(data);
+					var json = $.parseJSON(data);
 					if (json.success == false) {
 						alert(json.error);
 					}
@@ -72,7 +72,7 @@
 			
 			$.post(ajaxurl, params, function(data) {
 				try {
-					var json = eval(data);
+					var json = $.parseJSON(data);
 					if (json.success == false) {
 						alert(json.error);
 					}
